feat(input-handler): moving reading input from the main to separate file
